remove changed files only after sync

+++ client-side/lib-domain/src/main/java/com/tsourcecode/wiki/lib/domain/documents/staging/ChangedFilesController.kt
@@ -8,9 +8,13 @@ class ChangedFilesController(
         private val changedFilesDir: File,
         private val persistentStorageProvider: PersistentStorageProvider,
 ) {
-    private val storage = persistentStorageProvider.get("unstaged")
+    private val changedStorage = persistentStorageProvider.get("changed")
     private val changedFiles = mutableMapOf<String, String>().apply {
-        putAll(storage.all)
+        putAll(changedStorage.all)
+    }
+    private val stagedStorage = persistentStorageProvider.get("staged")
+    private val stagedFiles = mutableMapOf<String, String>().apply {
+        putAll(stagedStorage.all)
     }
 
     init {
@@ -20,18 +24,41 @@ class ChangedFilesController(
     fun markChanged(d: Document, modifiedContent: String) {
         val changedFile = d.toChangedFile()
         changedFiles[d.relativePath] = changedFile.absolutePath
-        storage.store(changedFiles)
+        changedStorage.store(changedFiles)
+
+        stagedFiles.remove(d.relativePath)
+        stagedStorage.store(stagedFiles)
         changedFile.writeText(modifiedContent)
     }
 
     fun markStaged(d: Document) {
+        val stagedFile = d.toChangedFile()
+        stagedFiles[d.relativePath] = stagedFile.absolutePath
+        stagedStorage.store(stagedFiles)
+    }
+
+    fun notifyFileSynced(d: Document) {
+        if (!changedFiles.contains(d.relativePath)) {
+            return
+        }
+
+        if (!stagedFiles.contains(d.relativePath)) {
+            return
+        }
+
         d.toChangedFile().delete()
         changedFiles.remove(d.relativePath)
-        storage.store(changedFiles)
+        stagedFiles.remove(d.relativePath)
+        changedStorage.store(changedFiles)
+        stagedStorage.store(changedFiles)
     }
 
     fun isChanged(d: Document): Boolean {
         return changedFiles.contains(d.relativePath)
+    }
+
+    fun isStaged(d: Document): Boolean {
+        return stagedFiles.contains(d.relativePath)
     }
 
     fun getChangedFile(d: Document): File? {
